Blocked out, attempted to implement android accelorometer controls - to no avail
