--- Roll-a-ballUnityProject/Assets/Controls.cs.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 78c73fab6792a204cb850467e5e34c69
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData:
